fix(ui): fold StreamHeader toolbar tests into existing suite
- Merge origin/main (brought in StreamHeader.vitest.mts from #8164) and
  fold StreamHeaderToolbar.vitest.mts's describe block into it as a
  sibling describe, per review-checklist §14 R7 (no duplicate suite for
  an already-covered component).
- Fix a latent bug this surfaced: two describe blocks in one file must
  share a single top-level useLitComponentTestDom() call, not one each —
  calling it per describe rebinds customElements to a second jsdom
  instance without stream-header registered, silently leaving
  element.shadowRoot undefined for the second block's tests.
- Add a comment on renderLabeledActionButtonParts noting it exists for
  API symmetry with renderIconActionButtonParts/renderLabeledActionButton
  rather than current caller count, per review suggestion.
